refactor: New application structure for pages
