+++ frontend/src/pages/App/stores/index.js
