+++ frontend/src/pages/Apps/stores/index.js
