+++ frontend/src/pages/Cli/stores/index.js
