+++ frontend/src/pages/Device/stores/index.js
