+++ frontend/src/pages/Device/stores/store-electron.js
