+++ frontend/src/pages/Device/stores/store-web.js
