+++ frontend/src/pages/InstalledApps/stores/index.js
