+++ frontend/src/pages/NfcTools/stores/store-electron.js
