+++ frontend/src/pages/Paint/stores/index.js
