+++ frontend/src/pages/Paint/stores/store-electron.js
