+++ frontend/src/pages/Paint/stores/store-web.js
