+++ frontend/src/pages/Pulseplot/index.vue
